pyproject.toml
[tool.poetry]
name = "websocket-stomp-client"
version = "0.1.0"
description = ""
authors = ["Luke Danielson <danielsonlukea@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
stomper = "^0.4.1"
websockets = "^7.0"
websocket-client = "^0.54.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.2"
poetry-setup = "^0.3.6"

[tool.poetry.scripts]
lint = "pylint websocket_stomp_client"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
